pyproject.toml
[project]
name = "findataagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "openai>=2.14.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "tushare>=1.4.24",
]
